// frontend/src/types.ts
export type ProductType = {
  _id: string;
  name: string;
  price: number;
  image: string;
  createdAt: string;
  updatedAt: string;
  __v: number;
}[];

export type InvidualProductType = {
  _id: string;
  name: string;
  price: number;
  image: string;
  createdAt: string;
  updatedAt: string;
  __v: number;
};